+++ src/Map/ObjectKeys.php
@@ -82,12 +82,6 @@ final class ObjectKeys implements Implementation
         return $this->values->count();
     }
 
-    #[\Override]
-    public function count(): int
-    {
-        return $this->size();
-    }
-
     /**
      * @param T $key
      *
@@ -332,7 +326,7 @@ final class ObjectKeys implements Implementation
          * @psalm-suppress MixedArgumentTypeCoercion
          * @psalm-suppress ImpureMethodCall
          */
-        $values->detach($key);
+        $values->offsetUnset($key);
         /** @psalm-suppress ImpureMethodCall */
         $values->rewind();
 
@@ -356,10 +350,10 @@ final class ObjectKeys implements Implementation
     /**
      * @param callable(T, S): bool $predicate
      *
-     * @return Map<bool, Map<T, S>>
+     * @return array{Map<T, S>, Map<T, S>}
      */
     #[\Override]
-    public function partition(callable $predicate): Map
+    public function partition(callable $predicate): array
     {
         $truthy = $this->clearMap();
         $falsy = $this->clearMap();
@@ -384,7 +378,7 @@ final class ObjectKeys implements Implementation
             }
         }
 
-        return Map::of([true, $truthy], [false, $falsy]);
+        return [$truthy, $falsy];
     }
 
     /**