+++ src/Map/Primitive.php
@@ -90,12 +90,6 @@ final class Primitive implements Implementation
         return \count($this->values);
     }
 
-    #[\Override]
-    public function count(): int
-    {
-        return $this->size();
-    }
-
     /**
      * @param T $key
      *
@@ -301,10 +295,10 @@ final class Primitive implements Implementation
     /**
      * @param callable(T, S): bool $predicate
      *
-     * @return Map<bool, Map<T, S>>
+     * @return array{Map<T, S>, Map<T, S>}
      */
     #[\Override]
-    public function partition(callable $predicate): Map
+    public function partition(callable $predicate): array
     {
         $truthy = $this->clearMap();
         $falsy = $this->clearMap();
@@ -320,7 +314,7 @@ final class Primitive implements Implementation
             }
         }
 
-        return Map::of([true, $truthy], [false, $falsy]);
+        return [$truthy, $falsy];
     }
 
     /**