+++ tests/MapTest.php
@@ -12,13 +12,6 @@ use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class MapTest extends TestCase
 {
-    public function testInterface()
-    {
-        $m = Map::of();
-
-        $this->assertInstanceOf(\Countable::class, $m);
-    }
-
     public function testOf()
     {
         $map = Map::of()
@@ -194,7 +187,7 @@ class MapTest extends TestCase
             ->put(3, 4);
         $count = 0;
 
-        $m->foreach(function(int $key, int $value) use (&$count) {
+        $_ = $m->foreach(function(int $key, int $value) use (&$count) {
             $this->assertSame($count, $key);
             $this->assertSame($value, $key + 1);
             ++$count;
@@ -368,31 +361,25 @@ class MapTest extends TestCase
             ->put(3, 4)
             ->put(4, 5);
 
-        $p = $m->partition(static function(int $i, int $v) {
+        [$true, $false] = $m->partition(static function(int $i, int $v) {
             return ($i + $v) % 3 === 0;
         });
 
-        $this->assertInstanceOf(Map::class, $p);
-        $this->assertNotSame($p, $m);
-        $this->assertSame(
-            [true, false],
-            $p->keys()->toList(),
-        );
         $this->assertSame(
             [1, 4],
-            $this->get($p, true)->keys()->toList(),
+            $true->keys()->toList(),
         );
         $this->assertSame(
             [2, 5],
-            $this->get($p, true)->values()->toList(),
+            $true->values()->toList(),
         );
         $this->assertSame(
             [0, 2, 3],
-            $this->get($p, false)->keys()->toList(),
+            $false->keys()->toList(),
         );
         $this->assertSame(
             [1, 3, 4],
-            $this->get($p, false)->values()->toList(),
+            $false->values()->toList(),
         );
     }