+++ tests/Set/LazyTest.php
@@ -31,7 +31,6 @@ class LazyTest extends TestCase
         });
 
         $this->assertSame(2, $set->size());
-        $this->assertSame(2, $set->count());
     }
 
     public function testIterator()
@@ -219,7 +218,7 @@ class LazyTest extends TestCase
 
         $this->assertSame([1, 2, 3, 4], $set->unsorted()->toList());
         $this->assertInstanceOf(Map::class, $groups);
-        $this->assertCount(2, $groups);
+        $this->assertSame(2, $groups->size());
         $this->assertSame([2, 4], $this->get($groups, 0)->toList());
         $this->assertSame([1, 3], $this->get($groups, 1)->toList());
     }
@@ -264,13 +263,11 @@ class LazyTest extends TestCase
             yield 3;
             yield 4;
         });
-        $groups = $set->partition(static fn($i) => $i % 2 === 0);
+        [$true, $false] = $set->partition(static fn($i) => $i % 2 === 0);
 
         $this->assertSame([1, 2, 3, 4], $set->unsorted()->toList());
-        $this->assertInstanceOf(Map::class, $groups);
-        $this->assertCount(2, $groups);
-        $this->assertSame([2, 4], $this->get($groups, true)->toList());
-        $this->assertSame([1, 3], $this->get($groups, false)->toList());
+        $this->assertSame([2, 4], $true->toList());
+        $this->assertSame([1, 3], $false->toList());
     }
 
     public function testSort()